+++ client/src/completion/DomainCompletionItemProvider.ts
@@ -4,26 +4,25 @@
  * ------------------------------------------------------------------------------------------ */
 'use strict';
 
-import { CompletionItem, TextDocument, Position, CompletionContext, CompletionItemKind, SnippetString, Range, CompletionTriggerKind, MarkdownString } from 'vscode';
+import { CompletionItem, TextDocument, Position, CompletionContext, CompletionItemKind, Range, MarkdownString } from 'vscode';
 import { PDDL } from '../../../common/src/parser';
 import { DomainInfo } from '../../../common/src/DomainInfo';
 import { PddlTokenType } from '../../../common/src/PddlTokenizer';
 import { PddlSyntaxNode } from '../../../common/src/PddlSyntaxNode';
 import { PddlStructure } from '../../../common/src/PddlStructure';
 import { nodeToRange } from '../utils';
+import { AbstractCompletionItemProvider, Suggestion } from './AbstractCompletionItemProvider';
 
-export class DomainCompletionItemProvider {
-
-    suggestionDetails: Map<string, SuggestionDetails>;
+export class DomainCompletionItemProvider extends AbstractCompletionItemProvider {
 
     constructor() {
-        this.suggestionDetails = new Map<string, SuggestionDetails>();
-        this.addSuggestionDocumentation(':requirements', 'Requirements', 'Required planning engine features.');
+        super();
+        this.addRequirementsDocumentation();
         this.addSuggestionDocumentation(':types', 'Types', new MarkdownString('Types of objects and their hierarchy. Example:').appendCodeblock('car - vehicle', PDDL));
         this.addSuggestionDocumentation(':constants', 'Constants', new MarkdownString('Constant objects that will be part of all problems defined for this domain in addition to the objects defined in the `:objects` section.'));
         this.addSuggestionDocumentation(':predicates', 'Predicates', 'Predicates are things that are either true or false.');
         this.addSuggestionDocumentation(':functions', 'Functions', 'Functions are used to define numeric values.');
-        this.addSuggestionDocumentation(':constraints', 'Constraints', 'Constraints.... you may want to stay away from those.');
+        this.addConstraintsDocumentation();
         this.addSuggestionDocumentation(':derived', 'Derived predicate/function', new MarkdownString('Derived predicate/function can be defined to simplify action declaration. Example derived predicate:').appendCodeblock('(:derived (p_and_q) \n\t(and (p) (q))\n)', PDDL).appendText('Example derived function: ').appendCodeblock('(:derived (c) (+ (a) (b))', PDDL));
         this.addSuggestionDocumentation(':action', 'Instantaneous action', new MarkdownString('Actions that change state of the world. Example:').appendCodeblock('(:action action_name\n\t:parameters (?t - type1)\n\t:precondition (and (p ?t))\n\t:effect (and (q ?t))\n)', PDDL));
         this.addSuggestionDocumentation(':durative-action', 'Durative action', 'Actions that change the state of the world when they start, then they last for a defined duration period, while changing the world continuously and finally change the state when they end.');
@@ -47,12 +46,16 @@ export class DomainCompletionItemProvider {
             '        (assign (velocity ?b) (* -0.8 (velocity ?b)))',
             '    )',
             ')'].join('\n'), PDDL).appendMarkdown('Note that `:process` and `:event` require the `:time` requirement.'),
-        CompletionItemKind.Event);
-    }
+            CompletionItemKind.Event);
 
-    addSuggestionDocumentation(label: string, detail: string, documentation: string | MarkdownString, kind?: CompletionItemKind) {
-        let details = new SuggestionDetails(label, detail, documentation, kind);
-        this.suggestionDetails.set(details.label, details);
+        /* ACTION KEYWORDS */
+        this.addSuggestionDocumentation(':parameters', 'Action parameters', new MarkdownString('Parameters such as:').appendCodeblock(':parameters (?v - vehicle ?from ?to - place)'));
+        this.addSuggestionDocumentation(':precondition', 'Instantaneous action precondition', '');
+        this.addSuggestionDocumentation(':effect', 'Action effect', '');
+
+        /* DURATIVE ACTION KEYWORDS */
+        this.addSuggestionDocumentation(':duration', 'Durative action duration', new MarkdownString('Examples:').appendCodeblock(':duration (= ?duration 1)\n:duration (and (>= ?duration (min_duration))(<= ?duration (max_duration)))'));
+        this.addSuggestionDocumentation(':condition', 'Durative action condition', '');
     }
 
     async provide(document: TextDocument, domainInfo: DomainInfo, position: Position, context: CompletionContext): Promise<CompletionItem[]> {
@@ -68,30 +71,59 @@ export class DomainCompletionItemProvider {
                 currentNode = currentNode.expand();
             }
 
-            let supportedSectionsHere = PddlStructure.getSupportedSectionsHere(currentNode, PddlStructure.PDDL_DOMAIN_SECTIONS, PddlStructure.PDDL_DOMAIN_STRUCTURES);
+            let supportedSectionsHere = PddlStructure.getSupportedSectionsHere(currentNode, currentNode, PddlTokenType.OpenBracketOperator, PddlStructure.PDDL_DOMAIN_SECTIONS, PddlStructure.PDDL_DOMAIN_STRUCTURES);
             let range = ['(', ':'].includes(context.triggerCharacter) ? nodeToRange(document, currentNode) : null;
 
-            let suggestions = supportedSectionsHere.map(s => Suggestion.from(s, context.triggerCharacter));
+            let suggestions = supportedSectionsHere.map(s => Suggestion.from(s, context.triggerCharacter, '('));
 
             return suggestions
-                .map((suggestion, index) => this.createCompletionItem(currentNode, suggestion, range, context, index))
+                .map((suggestion, index) => this.createDefineCompletionItem(currentNode, suggestion, range, context, index))
+                .filter(item => !!item); // filter out nulls
+        }
+        else if (this.insideAction(currentNode)) {
+            let nearestPrecedingKeyword = PddlStructure.getPrecedingKeywordOrSelf(currentNode);
+            let supportedSectionsHere = PddlStructure.getSupportedSectionsHere(nearestPrecedingKeyword, currentNode, PddlTokenType.Keyword, PddlStructure.PDDL_ACTION_SECTIONS, []);
+            let range = ['(', ':'].includes(context.triggerCharacter) ? nodeToRange(document, currentNode) : null;
+
+            let suggestions = supportedSectionsHere.map(s => Suggestion.from(s, context.triggerCharacter, ''));
+
+            return suggestions
+                .map((suggestion, index) => this.createActionCompletionItem(currentNode, suggestion, range, context, index))
+                .filter(item => !!item); // filter out nulls
+        }
+        else if (this.insideDurativeAction(currentNode)) {
+            let nearestPrecedingKeyword = PddlStructure.getPrecedingKeywordOrSelf(currentNode);
+            let supportedSectionsHere = PddlStructure.getSupportedSectionsHere(nearestPrecedingKeyword, currentNode, PddlTokenType.Keyword, PddlStructure.PDDL_DURATIVE_ACTION_SECTIONS, []);
+            let range = ['(', ':'].includes(context.triggerCharacter) ? nodeToRange(document, currentNode) : null;
+
+            let suggestions = supportedSectionsHere.map(s => Suggestion.from(s, context.triggerCharacter, ''));
+
+            return suggestions
+                .map((suggestion, index) => this.createDurativeActionCompletionItem(currentNode, suggestion, range, context, index))
                 .filter(item => !!item); // filter out nulls
         }
 
         return [];
     }
 
-    private insideDefine(domainInfo: DomainInfo, currentNode: PddlSyntaxNode, context: CompletionContext) {
-        let defineNode = domainInfo.syntaxTree.getDefineNode();
-        if (context.triggerKind === CompletionTriggerKind.Invoke) {
-            return currentNode.getParent() === defineNode;
-        }
-        else {
-            return currentNode.expand().getParent() === defineNode;
-        }
+    private insideAction(currentNode: PddlSyntaxNode): boolean {
+        const pattern = /^\(\s*:action$/i;
+        return this.insideScope(currentNode, pattern);
     }
 
-    createCompletionItem(_currentNode: PddlSyntaxNode, suggestion: Suggestion, range: Range, context: CompletionContext, index: number): CompletionItem | null {
+    private insideDurativeAction(currentNode: PddlSyntaxNode): boolean {
+        const pattern = /^\(\s*:durative-action$/i;
+        return this.insideScope(currentNode, pattern);
+    }
+
+    private insideScope(currentNode: PddlSyntaxNode, pattern: RegExp) {
+        let parentScope = currentNode.expand();
+        return parentScope.isType(PddlTokenType.OpenBracketOperator) && parentScope.getToken().tokenText.match(pattern) !== null;
+    }
+
+    private readonly DURATION_SNIPPET = ":duration ${2|(= ?duration 1),(> ?duration 0),(<= ?duration 10),(and (>= ?duration 1)(<= ?duration 2))|}";
+
+    createDefineCompletionItem(_currentNode: PddlSyntaxNode, suggestion: Suggestion, range: Range, context: CompletionContext, index: number): CompletionItem | null {
         if (!suggestion) { return null; }
         switch (suggestion.sectionName) {
             case PddlStructure.DOMAIN:
@@ -117,7 +149,7 @@ export class DomainCompletionItemProvider {
                 return this.createSnippetCompletionItem(suggestion, [
                     "(:durative-action ${1:action_name}",
                     "    :parameters ($0)",
-                    "    :duration ${2|(= ?duration 1),(> ?duration 0),(<= ?duration 10),(and (>= ?duration 1)(<= ?duration 2))|}",
+                    "    " + this.DURATION_SNIPPET,
                     "    :condition (and ",
                     "        (at start (and ",
                     "        ))",
@@ -164,40 +196,35 @@ export class DomainCompletionItemProvider {
         }
     }
 
-    private createSnippetCompletionItem(suggestion: Suggestion, snippet: string, range: Range, _context: CompletionContext, index: number) {
-        let suggestionDetail = this.suggestionDetails.get(suggestion.sectionName);
-        let completionItem = new CompletionItem(suggestion.sectionName, suggestionDetail.kind || CompletionItemKind.Keyword);
-        completionItem.insertText = new SnippetString(snippet);
-        if (range) { completionItem.range = range; }
-        if (suggestionDetail) {
-            completionItem.detail = suggestionDetail.detail;
-            completionItem.documentation = suggestionDetail.documentation;
-        }
-        completionItem.filterText = suggestion.filterText;
-        completionItem.sortText = 'item' + index;
-        return completionItem;
-    }
-}
-
-class Suggestion {
-    public readonly filterText: string;
-    constructor(public readonly sectionName: string) {
-        this.filterText = '(' + sectionName;
-    }
-
-    static from(sectionName: string, triggerCharacter: string): Suggestion | null {
-        if (triggerCharacter === ':' && !sectionName.startsWith(':')) {
-            return null;
-        }
-        else {
-            return new Suggestion(sectionName);
+    createActionCompletionItem(_currentNode: PddlSyntaxNode, suggestion: Suggestion, range: Range, context: CompletionContext, index: number): CompletionItem | null {
+        if (!suggestion) { return null; }
+        switch (suggestion.sectionName) {
+            case PddlStructure.PARAMETERS:
+                return this.createParametersCompletionItem(suggestion, range, context, index);
+            case PddlStructure.PRECONDITION:
+            case PddlStructure.EFFECT:
+                return this.createSnippetCompletionItem(suggestion, suggestion.sectionName + " (and \n\t$0\n)", range, context, index);
+            default:
+                return null;
         }
     }
-}
 
-class SuggestionDetails {
-    
-    constructor(public readonly label: string, public readonly detail: string, public readonly documentation: string | MarkdownString, public readonly kind?: CompletionItemKind) {
+    private createParametersCompletionItem(suggestion: Suggestion, range: Range, context: CompletionContext, index: number): CompletionItem {
+        return this.createSnippetCompletionItem(suggestion, ":parameters ($0)", range, context, index);
+    }
 
+    createDurativeActionCompletionItem(_currentNode: PddlSyntaxNode, suggestion: Suggestion, range: Range, context: CompletionContext, index: number): CompletionItem | null {
+        if (!suggestion) { return null; }
+        switch (suggestion.sectionName) {
+            case PddlStructure.PARAMETERS:
+                return this.createParametersCompletionItem(suggestion, range, context, index);
+            case PddlStructure.DURATION:
+                return this.createSnippetCompletionItem(suggestion, this.DURATION_SNIPPET, range, context, index);
+            case PddlStructure.CONDITION:
+            case PddlStructure.EFFECT:
+                return this.createSnippetCompletionItem(suggestion, suggestion.sectionName + " (and \n\t$0\n)", range, context, index);
+            default:
+                return null;
+        }
     }
 }